# Swift Breakthrough: vectorized