--- a/.gitignore
+++ b/.gitignore
@@ -23,6 +23,9 @@ cmake-build-*/
 dist/
 pyspiel.egg-info/
 
+# Swift build directory
+.build
+
  # External git modules
 open_spiel/abseil-cpp/
 open_spiel/games/bridge/double_dummy_solver/
@@ -38,8 +41,11 @@ open_spiel/scripts/jill.sh
 # julia wrapper
 Manifest.toml
 
+
 # IDE
 .idea/
 .vscode/
+*~
+
 
 open_spiel/cmake-build-debug/