--- a/swift/Benchmarks/main.swift
+++ b/swift/Benchmarks/main.swift
@@ -33,6 +33,10 @@ benchmark("random game: Breakthrough") {
   playRandomGame(Breakthrough())
 }
 
+benchmark("random game: FastBreakthrough") {
+  playRandomGame(FastBreakthrough())
+}
+
 benchmark("random game: Texas Hold'em") {
   playRandomGame(TexasHoldem(playerCount: 3))
 }