… swift/Sources/OpenSpiel/FastBreakthrough.swift