After #230, we were still spending
a significant amount of time inside the action getter. I had a hunch that
we could optimize the inner loops. This change splits the legal action getter
into two components. The first is a vectorized operation over all board
board positions. The second is a set of loops to convert the bitmasks into
actions. This results in another ~1.5x speedup.

Note: because I haven't made the bitmask operations general (and that only
work on a fixed 8x8 (the most common size) board), I have made this a
separate implementation `FastBreakthrough`. This also makes for convenient
side-by-side benchmark comparisons:

```
name                          time        std        iterations
---------------------------------------------------------------
random game: TicTacToe         11349.0 ns ±  15.51 %     127874
random game: Kuhn Poker         4836.0 ns ±  12.11 %     273561
random game: Breakthrough     100297.0 ns ±  24.83 %      13980
random game: FastBreakthrough  67811.0 ns ±  25.00 %      20697
random game: Texas Hold'em     71461.5 ns ± 665.52 %       5436
```

I have some suspicions that simplifying the Action type might result in
further speedups. Based on a quick profile, we still spend ~30% of the
time inside the conversion from masks to the Action struct. (By contrast,
[C++ uses a simple Int64](https://github.com/deepmind/open_spiel/blob/70f6c2fd9ab964e25e5b0151b6a7e47a5855a2af/open_spiel/spiel_utils.h#L103).
I would just like to note here that this change further makes this an
Apples-to-Oranges performance comparison. At this point, the algorithms
are different, and the API exposed is also quite different.)

Related issue: #228